refactor: Header 컴포넌트 위치 'layout' 폴더로 이동
