+++ src/components/layout/Header/Header.jsx
@@ -1,8 +1,8 @@
 /** @jsxImportSource @emotion/react */
 import { css } from "@emotion/react";
-import logoPandaImg from "../../assets/images/logo-panda.svg";
-import logoTxtImg from "../../assets/images/logo-txt.svg";
-import avatarImg from "../../assets/images/img-avatar.png";
+import logoPandaImg from "../../../assets/images/logo-panda.svg";
+import logoTxtImg from "../../../assets/images/logo-txt.svg";
+import avatarImg from "../../../assets/images/img-avatar.png";
 import Nav from "../../Nav";
 import Avatar from "../../Avatar";
